+++ extensions/coffee/src/caffeinate.ts
@@ -1,7 +1,7 @@
-import { startCaffeinate, getSchedule, changeScheduleState } from "./utils";
+import { startCaffeinate, getSchedule, changeScheduleState, deviceName } from "./utils";
 
 export default async () => {
   const schedule = await getSchedule();
   if (schedule != undefined) await changeScheduleState("decaffeinate", schedule);
-  await startCaffeinate({ menubar: true, status: true }, "Your Mac is now caffeinated");
+  await startCaffeinate({ menubar: true, status: true }, `Your ${deviceName()} is now caffeinated`);
 };